コンテンツ系のuseAsyncData問題を修正
