+++ components/base/overlay.vue
@@ -34,7 +34,6 @@ withDefaults(
   top: 0;
   width: 100%;
   height: 100%;
-  background-color: rgba(0, 0, 0, 0.25);
 }
 .overlay-now-buzy {
   display: flex;
