+++ composables/use-content/use-eyecatch-form.ts
@@ -1,0 +1,61 @@
+import { useForm, useField } from 'vee-validate'
+import type { EyecatchType, EyecatchForm } from '@/types/content'
+
+/**
+ * Eyecatch Form
+ */
+export const useEyecatchForm = () => {
+  const { noBlank, maxLength } = useValidateRules()
+  const eyecatchFormSchema = {
+    title: (v: string | undefined) => {
+      if (!noBlank(v)) return 'トップタイトルを入力してください'
+      if (!maxLength(v, 40)) return '40文字以内で入力してください'
+      return true
+    },
+    subtitle: (v: string | undefined) =>
+      maxLength(v, 50) || '50文字以内で入力してください',
+    image: (v: string | undefined) =>
+      noBlank(v) || 'トップ背景画像ファイルを設定してください',
+    imageFile: () => true,
+  }
+  const eyecatchFormInitial: EyecatchForm = {
+    title: '',
+    subtitle: '',
+    image: '',
+    imageFile: null,
+  }
+
+  const { handleSubmit, handleReset, validate } = useForm({
+    validationSchema: eyecatchFormSchema,
+    initialValues: eyecatchFormInitial,
+  })
+
+  const formData = {
+    title: useField<EyecatchForm['title']>('title'),
+    subtitle: useField<EyecatchForm['subtitle']>('subtitle'),
+    image: useField<EyecatchForm['image']>('image'),
+    imageFile: useField<EyecatchForm['imageFile']>('imageFile'),
+  }
+
+  const resetEyeCatchForm = (eyecatchData?: EyecatchType | null) => {
+    if (!eyecatchData) return
+    formData.title.value.value = eyecatchData?.title ?? ''
+    formData.subtitle.value.value = eyecatchData?.subtitle ?? ''
+    formData.image.value.value = eyecatchData?.image?.url ?? ''
+    formData.imageFile.value.value = null
+  }
+
+  const changeImageFile = (params: { file: File; url: string }) => {
+    formData.image.value.value = params.url
+    formData.imageFile.value.value = params.file
+  }
+
+  return {
+    handleSubmit,
+    handleReset,
+    validate,
+    formData,
+    resetEyeCatchForm,
+    changeImageFile,
+  }
+}
