+++ composables/use-content/use-information.ts
@@ -1,0 +1,175 @@
+import { useForm, useField } from 'vee-validate'
+import type {
+  InformationType,
+  InformationForm,
+  ImageSettings,
+} from '@/types/content'
+import type { InformationGetApi, InformationSaveApi } from '@/types/content-api'
+
+const apiUrl = '/informations'
+const useInformationContent = (customerId: Ref<number | null>) => {
+  const {
+    loadData,
+    get,
+    setImageSettings,
+    contentDataRef,
+    loadingRef: readLoading,
+  } = useContentRead<InformationGetApi>(customerId, apiUrl)
+  const {
+    create,
+    update,
+    remove,
+    updateImageSettingsWithDebounced,
+    loadingRef: writeLoading,
+  } = useContentWrite<InformationSaveApi, InformationGetApi>(customerId, apiUrl)
+
+  const apitypeToInformationType = (
+    apiData?: InformationGetApi | null
+  ): InformationType | null =>
+    apiData
+      ? {
+          id: apiData.id,
+          customerId: apiData.customerId,
+          title: apiData.title,
+          subtitle: apiData.subtitle,
+          body: apiData.body,
+          image: apiData.image,
+        }
+      : null
+
+  const informationRef = computed<InformationType | null>(() =>
+    apitypeToInformationType(contentDataRef.value)
+  )
+  const loading = computed(() => readLoading.value || writeLoading.value)
+
+  const setInformationImageSettings = (
+    settings: Partial<ImageSettings>
+  ): ImageSettings | void => {
+    if (!informationRef.value) return
+    if (!informationRef.value.image?.settings) return
+
+    const newSettings: ImageSettings = {
+      ...informationRef.value.image.settings,
+      ...settings,
+    }
+    setImageSettings(newSettings)
+    return newSettings
+  }
+
+  const createInformation = async (
+    formData: InformationForm
+  ): Promise<InformationType | null> => {
+    const inputData: InformationSaveApi = {
+      customerId: customerId.value ?? 0,
+      title: formData.title,
+      subtitle: formData.subtitle,
+      body: formData.body,
+      imageFile: formData.imageFile,
+    }
+    const data = await create(inputData)
+    return apitypeToInformationType(data ?? null)
+  }
+
+  const updateInformation = async (
+    contentId: number,
+    formData: InformationForm
+  ): Promise<InformationType | null> => {
+    const inputData: InformationSaveApi = {
+      customerId: customerId.value ?? 0,
+      title: formData.title,
+      subtitle: formData.subtitle,
+      body: formData.body,
+      imageFile: formData.imageFile,
+    }
+    const data = await update(contentId, inputData)
+    return apitypeToInformationType(data ?? null)
+  }
+
+  const updateInformationImageSettings = (
+    contentId: number,
+    imageSettings: ImageSettings
+  ) => {
+    updateImageSettingsWithDebounced(contentId, imageSettings)
+  }
+
+  return {
+    loadInformation: loadData,
+    getInformation: get,
+    setInformationImageSettings,
+    createInformation,
+    updateInformation,
+    removeInformation: remove,
+    updateInformationImageSettings,
+    informationRef,
+    loading,
+  }
+}
+
+/**
+ * Information API アクションサービス
+ * @param customerId
+ */
+export const useInformationActions = (customerId: Ref<number | null>) => {
+  const {
+    loadInformation,
+    getInformation,
+    setInformationImageSettings,
+    createInformation,
+    updateInformation,
+    removeInformation,
+    updateInformationImageSettings,
+    informationRef,
+    loading,
+  } = useInformationContent(customerId)
+
+  const { addSnackber } = useSnackbars()
+
+  const onLoad = async () => {
+    await loadInformation()
+  }
+
+  const onCreate = async (formData: InformationForm) => {
+    const savedData = await createInformation(formData)
+    addSnackber?.('Information を登録しました。')
+    getInformation(savedData?.id)
+  }
+
+  const onUpdate = async ({
+    id,
+    formData,
+  }: {
+    id: number
+    formData: InformationForm
+  }) => {
+    if (!id) return
+
+    const savedData = await updateInformation(id, formData)
+    addSnackber?.('Information を更新しました。')
+    getInformation(savedData?.id)
+  }
+
+  const onRemove = async (id: number) => {
+    await removeInformation(id)
+    addSnackber?.('Information を削除しました。')
+    getInformation()
+  }
+
+  const onUpdateImageSetting = (settings: Partial<ImageSettings>) => {
+    if (!informationRef.value?.id) return
+
+    const newSettings = setInformationImageSettings(settings)
+    if (!newSettings) return
+
+    updateInformationImageSettings(informationRef.value.id, newSettings)
+  }
+
+  return {
+    informationRef,
+    onLoad,
+    onCreate,
+    onUpdate,
+    onRemove,
+    onUpdateImageSetting,
+    loading,
+  }
+}
