+++ types/content.d.ts
@@ -1,4 +1,4 @@
-import type { NewsCategory2Label } from '@/composables/use-news-category'
+import type { NewsCategory2Label } from '@/types/news-category'
 
 /**
  * Image Data Types
